pyproject.toml
[project]
name = "Distill"
version = "0.1.0"
description = "LLM Prompt Compression Library"
requires-python = ">=3.14"
dependencies = [
    "datasets>=4.5.0",
    "fastapi[standard]>=0.128.0",
    "loguru>=0.7.3",
    "nltk>=3.9.2",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
    "tokenc>=0.1.2",
    "torch>=2.9.1",
    "transformers>=4.57.6",
]
